make utils header-only
