+++ src/ecos/util/uuid.hpp
@@ -2,12 +2,46 @@
 #ifndef ECOS_UUID_HPP
 #define ECOS_UUID_HPP
 
+#include <ios>
+#include <random>
+#include <sstream>
 #include <string>
 
 namespace ecos
 {
 
-std::string generate_uuid();
+inline std::string generate_uuid()
+{
+    static std::random_device rd;
+    static std::mt19937 gen(rd());
+    static std::uniform_int_distribution<> dis(0, 15);
+    static std::uniform_int_distribution<> dis2(8, 11);
+
+    int i;
+    std::stringstream ss;
+    ss << std::hex;
+    for (i = 0; i < 8; i++) {
+        ss << dis(gen);
+    }
+    ss << "-";
+    for (i = 0; i < 4; i++) {
+        ss << dis(gen);
+    }
+    ss << "-4";
+    for (i = 0; i < 3; i++) {
+        ss << dis(gen);
+    }
+    ss << "-";
+    ss << dis2(gen);
+    for (i = 0; i < 3; i++) {
+        ss << dis(gen);
+    }
+    ss << "-";
+    for (i = 0; i < 12; i++) {
+        ss << dis(gen);
+    }
+    return ss.str();
+}
 
 } // namespace ecos
